fix-loign bug
