+++ app/build.gradle.kts
@@ -12,7 +12,7 @@ android {
         minSdk = 26
         targetSdk = 35
 
-        versionCode = 31
+        versionCode = 32
 
         versionName = "1.0.2"
