+++ app/src/main/java/com/slabstech/dhwani/voiceai/Api.kt
@@ -1,20 +1,11 @@
 package com.slabstech.dhwani.voiceai
 
-import android.app.Application
-import android.content.Context
-import android.util.Log
-import androidx.preference.PreferenceManager
-import kotlinx.coroutines.runBlocking
 import okhttp3.MultipartBody
-import okhttp3.OkHttpClient
 import okhttp3.RequestBody
 import okhttp3.ResponseBody
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.*
-import java.util.concurrent.TimeUnit
 
+// Data classes
 data class LoginRequest(val username: String, val password: String)
 data class TokenResponse(val access_token: String, val token_type: String)
 data class TranscriptionRequest(val language: String)
@@ -25,6 +16,7 @@ data class TranslationRequest(val sentences: List<String>, val src_lang: String,
 data class TranslationResponse(val translations: List<String>)
 data class VisualQueryResponse(val answer: String)
 
+// API Service Interface
 interface ApiService {
     @POST("v1/token")
     suspend fun login(@Body loginRequest: LoginRequest): TokenResponse
@@ -71,55 +63,4 @@ interface ApiService {
 
     @POST("v1/refresh")
     suspend fun refreshToken(@Header("Authorization") token: String): TokenResponse
-}
-
-object RetrofitClient {
-    private const val BASE_URL_DEFAULT = "https://slabstech-dhwani-server.hf.space/"
-
-    private fun getOkHttpClient(context: Context): OkHttpClient {
-        val prefs = PreferenceManager.getDefaultSharedPreferences(context)
-        val authenticator = object : okhttp3.Authenticator {
-            override fun authenticate(route: okhttp3.Route?, response: okhttp3.Response): okhttp3.Request? {
-                val currentToken = prefs.getString("access_token", null) ?: return null
-                try {
-                    val refreshResponse = runBlocking {
-                        apiService(context).refreshToken("Bearer $currentToken")
-                    }
-                    val newToken = refreshResponse.access_token
-                    prefs.edit().putString("access_token", newToken).apply()
-                    return response.request.newBuilder()
-                        .header("Authorization", "Bearer $newToken")
-                        .build()
-                } catch (e: Exception) {
-                    Log.e("RetrofitClient", "Token refresh failed: ${e.message}", e)
-                    return null
-                }
-            }
-        }
-
-        return OkHttpClient.Builder()
-            .authenticator(authenticator)
-            .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-            .connectTimeout(30, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS)
-            .writeTimeout(30, TimeUnit.SECONDS)
-            .build()
-    }
-
-    fun apiService(context: Context): ApiService {
-        val prefs = PreferenceManager.getDefaultSharedPreferences(context)
-        val baseUrl = prefs.getString("base_url", BASE_URL_DEFAULT) ?: BASE_URL_DEFAULT
-        return Retrofit.Builder()
-            .baseUrl(baseUrl)
-            .client(getOkHttpClient(context))
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(ApiService::class.java)
-    }
-}
-
-class DhwaniApp : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
 }
